Fix overflow in reweighting by using long unsigned ints

+++ include/reweighting.h
@@ -1,2 +1,2 @@
-void calculate_weights(conjugrad_float_t *w, unsigned char *msa, int ncol, int nrow, conjugrad_float_t threshold);
-void uniform_weights(conjugrad_float_t *w, int nrow);
+void calculate_weights(conjugrad_float_t *w, unsigned char *msa, uint64_t ncol, uint64_t nrow, conjugrad_float_t threshold);
+void uniform_weights(conjugrad_float_t *w, uint64_t nrow);
